added API endpoint integration chanegs and created js file for pricing-page component
